% Generated by roxygen2 (4.0.1): do not edit by hand
\name{format_iso_8601}
\alias{format_iso_8601}
\title{Format date and time according to ISO 8601}
\usage{
format_iso_8601(date, tz = "UTC")
}
\arguments{
\item{date}{The date(s) for format}

\item{tz}{Time zone.}
}
\description{
Format date and time according to ISO 8601
}

