% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_iso_8601}
\alias{parse_iso_8601}
\title{Parse date from an ISO 8601 format}
\usage{
parse_iso_8601(dates)
}
\arguments{
\item{dates}{A character vector containing the dates}
}
\value{
A \code{POSIXct} vector.
}
\description{
Parse date from an ISO 8601 format
}

