% Generated by roxygen2 (4.0.1): do not edit by hand
\name{format_iso_8601}
\alias{format_iso_8601}
\title{Format date and time according to ISO 8601}
\usage{
format_iso_8601(date)
}
\arguments{
\item{date}{The date(s) to format.}
}
\value{
Character vector of formatted dates.
}
\description{
Format a date in a fixed format that is ISO 8601 valid, and
can be used to compare dates as character strings. It converts
the date(s) to UTC.
}
\examples{
format_iso_8601(parse_iso_8601("2013-02-08"))
format_iso_8601(parse_iso_8601("2013-02-08 09:34:00"))
format_iso_8601(parse_iso_8601("2013-02-08 09:34:00+01:00"))
format_iso_8601(parse_iso_8601("2013-W06-5"))
format_iso_8601(parse_iso_8601("2013-039"))
}

